update migration for new datamodel

+++ database/migrations/2021_02_02_121148_create_rooms_table.php
@@ -22,7 +22,7 @@ class CreateRoomsTable extends Migration
                 2,
                 true
             );
-            $table->tinyInteger('beds', false, true);
+            $table->smallInteger('beds', false, true);
             $table->text('description');
             $table->string('image_path_1');
             $table->string('image_path_2')->nullable();
